+++ src/tracer/ptrace.rs
@@ -173,7 +173,7 @@ pub fn attach_all_threads(pid: Pid) -> Result<(Vec<Thread>, usize)> {
 impl Drop for Thread {
     fn drop(&mut self) {
         if let Err(e) = ptrace::detach(self.tid, None) {
-            log::warn!("Cannot ptrace::detach from {}", self.tid);
+            log::warn!("Cannot ptrace::detach from {}: {}", self.tid, e);
         }
     }
 }
